backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docstring_parser==0.16
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
idna==3.10
numpy==2.2.5
packaging==25.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9.1
shapely==2.1.0
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
vertexai==1.71.1
